fix(models): 주차장 시간 필드 분수값 허용 (basicTime/addUnitTime int→float)
Fixes #6

전국주차장정보표준데이터 live 응답에 addUnitTime='0.5' 분수 row가 존재 —
int 선언이라 pydantic int_parsing ValidationError로 페이지 전체가 거부됐다
(python-krtour-map T-212e run ea127324). basic_time도 같은 시간 단위라 함께
float로 완화. 회귀 테스트 추가.

검증: pytest tests/unit 15 passed.

+++ src/datagokr/models.py
@@ -126,9 +126,10 @@ class PublicParkingLot(StandardItem):
     holiday_oper_open_hhmm: str | None = Field(default=None, alias="holidayOperOpenHhmm")
     holiday_close_open_hhmm: str | None = Field(default=None, alias="holidayCloseOpenHhmm")
     parkingchrge_info: str | None = Field(default=None, alias="parkingchrgeInfo")
-    basic_time: int | None = Field(default=None, alias="basicTime")
+    # 시간 필드는 live 데이터에 분수값이 존재한다 (예: addUnitTime='0.5' — #6).
+    basic_time: float | None = Field(default=None, alias="basicTime")
     basic_charge: int | None = Field(default=None, alias="basicCharge")
-    add_unit_time: int | None = Field(default=None, alias="addUnitTime")
+    add_unit_time: float | None = Field(default=None, alias="addUnitTime")
     add_unit_charge: int | None = Field(default=None, alias="addUnitCharge")
     day_cmmtkt_adj_time: float | None = Field(default=None, alias="dayCmmtktAdjTime")
     day_cmmtkt: int | None = Field(default=None, alias="dayCmmtkt")
